requirements.txt
pika
pymongo
netmiko
pytest-dotenv
fastapi[all]
black
flake8
